Silence remaining compiler warnings under Solaris 8 (save for the utmp code, bug Bug 1824 will address that).
The nlink_t/pid_t casts will be unnecessary once the Solaris _LP64 ABI is
the default (see sys/isa_defs.h for more info on the Solaris ABIs). I didn't
want to move from _ILP32 (the historical default) to _LP64 because I wasn't
sure of the ramifications. We don't have any DSOs to worry about, but I'm
not sure what else would change/break.

+++ modules/mod_ls.c
@@ -25,7 +25,7 @@
  */
 
 /* Directory listing module for ProFTPD.
- * $Id: mod_ls.c,v 1.66 2002-10-21 17:06:09 castaglia Exp $
+ * $Id: mod_ls.c,v 1.67 2002-11-23 18:39:31 jwm Exp $
  */
 
 #include "conf.h"
@@ -414,18 +414,20 @@ static int listfile(cmd_rec *cmd, pool *p, const char *name) {
         if (!opt_n) {
 
           /* Format nameline using user/group names. */
-          snprintf(nameline, sizeof(nameline), "%s %3d %-8s %-8s %8" PR_LU
-            " %s %2d %s %s", m, st.st_nlink, MAP_UID(st.st_uid),
-            MAP_GID(st.st_gid), st.st_size, months[t->tm_mon], t->tm_mday,
-            timeline, name);
+          snprintf(nameline, sizeof(nameline),
+                   "%s %3d %-8s %-8s %8" PR_LU " %s %2d %s %s",
+                   m, (int)st.st_nlink,
+                   MAP_UID(st.st_uid), MAP_GID(st.st_gid), st.st_size,
+                   months[t->tm_mon], t->tm_mday, timeline, name);
 
         } else {
 
           /* Format nameline using user/group IDs. */
-          snprintf(nameline, sizeof(nameline), "%s %3d %-8u %-8u %8" PR_LU
-            " %s %2d %s %s", m, st.st_nlink, st.st_uid, st.st_gid,
-            st.st_size, months[t->tm_mon], t->tm_mday,
-            timeline, name);
+          snprintf(nameline, sizeof(nameline),
+                   "%s %3d %-8u %-8u %8" PR_LU " %s %2d %s %s",
+                   m, (int)st.st_nlink,
+                   (unsigned)st.st_uid, (unsigned)st.st_gid, st.st_size,
+                   months[t->tm_mon], t->tm_mday, timeline, name);
         }
 
         if (S_ISLNK(st.st_mode)) {
